Added new plugin - prioritize
