+++ pkg/scheduler/actions/preempt/preempt.go
@@ -175,14 +175,29 @@ func preempt(
 ) (bool, error) {
 	resreq := preemptor.Resreq.Clone()
 	preempted := api.EmptyResource()
-
+	predicateNodes := []*api.NodeInfo{}
+	nodeScores := map[int][]*api.NodeInfo{}
 	assigned := false
 
 	for _, node := range nodes {
 		if err := ssn.PredicateFn(preemptor, node); err != nil {
+			glog.V(3).Infof("Predicates failed for task <%s/%s> on node <%s>: %v",
+				preemptor.Namespace, preemptor.Name, node.Name, err)
 			continue
+		} else {
+			predicateNodes = append(predicateNodes, node)
 		}
-
+	}
+	for _, node := range predicateNodes {
+		score, err := ssn.PriorityFn(preemptor, node)
+		if err != nil {
+			glog.V(3).Infof("Error in Calculating Priority for the node:%v", err)
+		} else {
+			nodeScores[score] = append(nodeScores[score], node)
+		}
+	}
+	selectedNodes := util.SelectBestNode(nodeScores)
+	for _, node := range selectedNodes {
 		glog.V(3).Infof("Considering Task <%s/%s> on Node <%s>.",
 			preemptor.Namespace, preemptor.Name, node.Name)
