+++ pkg/scheduler/factory.go
@@ -25,6 +25,7 @@ import (
 	"github.com/kubernetes-sigs/kube-batch/pkg/scheduler/plugins/drf"
 	"github.com/kubernetes-sigs/kube-batch/pkg/scheduler/plugins/gang"
 	"github.com/kubernetes-sigs/kube-batch/pkg/scheduler/plugins/predicates"
+	"github.com/kubernetes-sigs/kube-batch/pkg/scheduler/plugins/prioritize"
 	"github.com/kubernetes-sigs/kube-batch/pkg/scheduler/plugins/priority"
 	"github.com/kubernetes-sigs/kube-batch/pkg/scheduler/plugins/proportion"
 
@@ -37,6 +38,7 @@ func init() {
 	framework.RegisterPluginBuilder("gang", gang.New)
 	framework.RegisterPluginBuilder("predicates", predicates.New)
 	framework.RegisterPluginBuilder("priority", priority.New)
+	framework.RegisterPluginBuilder("prioritize", prioritize.New)
 
 	// Plugins for Queues
 	framework.RegisterPluginBuilder("proportion", proportion.New)
