+++ pkg/scheduler/plugins/priority/priority.go
@@ -17,102 +17,12 @@ limitations under the License.
 package priority
 
 import (
-	"fmt"
 	"github.com/golang/glog"
 	"github.com/kubernetes-sigs/kube-batch/pkg/scheduler/api"
 	"github.com/kubernetes-sigs/kube-batch/pkg/scheduler/framework"
-	priority "github.com/kubernetes-sigs/kube-batch/pkg/scheduler/util/priorities"
-	"k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/kubernetes/pkg/scheduler/algorithm"
-	"k8s.io/kubernetes/pkg/scheduler/algorithm/priorities"
-	"k8s.io/kubernetes/pkg/scheduler/cache"
 )
 
 type priorityPlugin struct {
-}
-
-func generateNodeMapAndSlice(nodes []*api.NodeInfo) (map[string]*cache.NodeInfo, []*v1.Node) {
-	var nodeMap map[string]*cache.NodeInfo
-	var nodeSlice []*v1.Node
-	for _, node := range nodes {
-		nodeInfo := cache.NewNodeInfo(node.Pods()...)
-		nodeInfo.SetNode(node.Node)
-		nodeMap[node.Name] = nodeInfo
-		nodeSlice = append(nodeSlice, node.Node)
-	}
-	return nodeMap, nodeSlice
-}
-
-type cachedNodeInfo struct {
-	session *framework.Session
-}
-
-func (c *cachedNodeInfo) GetNodeInfo(name string) (*v1.Node, error) {
-	node, found := c.session.NodeIndex[name]
-	if !found {
-		return nil, fmt.Errorf("failed to find node <%s>", name)
-	}
-
-	return node.Node, nil
-}
-
-type podLister struct {
-	session *framework.Session
-}
-
-func (pl *podLister) List(selector labels.Selector) ([]*v1.Pod, error) {
-	var pods []*v1.Pod
-	for _, job := range pl.session.Jobs {
-		for status, tasks := range job.TaskStatusIndex {
-			if !api.AllocatedStatus(status) {
-				continue
-			}
-
-			for _, task := range tasks {
-				if selector.Matches(labels.Set(task.Pod.Labels)) {
-					pod := task.Pod.DeepCopy()
-					pod.Spec.NodeName = task.NodeName
-					pods = append(pods, pod)
-				}
-			}
-		}
-	}
-
-	return pods, nil
-}
-
-func (pl *podLister) FilteredList(podFilter algorithm.PodFilter, selector labels.Selector) ([]*v1.Pod, error) {
-	var pods []*v1.Pod
-	for _, job := range pl.session.Jobs {
-		for status, tasks := range job.TaskStatusIndex {
-			if !api.AllocatedStatus(status) {
-				continue
-			}
-
-			for _, task := range tasks {
-				if podFilter(task.Pod) && selector.Matches(labels.Set(task.Pod.Labels)) {
-					pod := task.Pod.DeepCopy()
-					pod.Spec.NodeName = task.NodeName
-					pods = append(pods, pod)
-				}
-			}
-		}
-	}
-
-	return pods, nil
-}
-
-type nodeLister struct {
-	session *framework.Session
-}
-
-func (nl *nodeLister) List() ([]*v1.Node, error) {
-	var nodes []*v1.Node
-	for _, node := range nl.session.Nodes {
-		nodes = append(nodes, node.Node)
-	}
-	return nodes, nil
 }
 
 func New() framework.Plugin {
@@ -164,82 +74,6 @@ func (pp *priorityPlugin) OnSessionOpen(ssn *framework.Session) {
 	}
 
 	ssn.AddJobOrderFn(pp.Name(), jobOrderFn)
-
-	priorityFn := func(task *api.TaskInfo, node *api.NodeInfo, nodes []*api.NodeInfo) (int, error) {
-		nodeInfo := cache.NewNodeInfo(node.Pods()...)
-		nodeInfo.SetNode(node.Node)
-		var score int = 0
-
-		host, err := priority.ImageLocalityPriorityMap(task.Pod, len(nodes), nodeInfo)
-		if err != nil {
-			glog.V(3).Infof("Image Locality Priority Failed because of Error: %v", err)
-			return 0, err
-		}
-		score = score + host.Score
-
-		host, err = priorities.BalancedResourceAllocationMap(task.Pod, nil, nodeInfo)
-		if err != nil {
-			glog.V(3).Infof("Balanced Resource Allocation Priority Failed because of Error: %v", err)
-			return 0, err
-		}
-		score = score + host.Score
-
-		host, err = priorities.LeastRequestedPriorityMap(task.Pod, nil, nodeInfo)
-		if err != nil {
-			glog.V(3).Infof("Least Requested Priority Failed because of Error: %v", err)
-			return 0, err
-		}
-		score = score + host.Score
-
-		host, err = priorities.MostRequestedPriorityMap(task.Pod, nil, nodeInfo)
-		if err != nil {
-			glog.V(3).Infof("Most Requested Priority Failed because of Error: %v", err)
-			return 0, err
-		}
-		score = score + host.Score
-
-		host, err = priorities.CalculateNodePreferAvoidPodsPriorityMap(task.Pod, nil, nodeInfo)
-		if err != nil {
-			glog.V(3).Infof("Calculate Node Prefer Avoid Pods Priority Failed because of Error: %v", err)
-			return 0, err
-		}
-		score = score + host.Score
-
-		host, err = priorities.ResourceLimitsPriorityMap(task.Pod, nil, nodeInfo)
-		if err != nil {
-			glog.V(3).Infof("Resource Limits Priority Failed because of Error: %v", err)
-			return 0, err
-		}
-		score = score + host.Score
-
-		host, err = priorities.ComputeTaintTolerationPriorityMap(task.Pod, nil, nodeInfo)
-		if err != nil {
-			glog.V(3).Infof("Compute Taint Toleration Priority Failed because of Error: %v", err)
-			return 0, err
-		}
-		score = score + host.Score
-
-		host, err = priorities.CalculateNodeAffinityPriorityMap(task.Pod, nil, nodeInfo)
-		if err != nil {
-			glog.V(3).Infof("Calculate Node Affinity Priority Failed because of Error: %v", err)
-			return 0, err
-		}
-		score = score + host.Score
-
-		for label, _ := range task.Pod.Labels {
-			mapFn, _ := priorities.NewNodeLabelPriority(label, true)
-			host, err := mapFn(task.Pod, nil, nodeInfo)
-			if err != nil {
-				glog.V(3).Infof("Calculate Node Label Priority Failed because of Error: %v", err)
-				return 0, err
-			}
-			score = score + host.Score
-		}
-
-		glog.V(3).Infof("Total Score for that node is: %f", score)
-		return score, nil
-	}
-	ssn.AddPriorityFn(pp.Name(), priorityFn)
 }
 
 func (pp *priorityPlugin) OnSessionClose(ssn *framework.Session) {}
